replace sleep with verify_record
instead of **hoping** that 10 seconds is enough for the DNS record to propagate, instead use DNS to see if it's propagated. Only release the flow if it has finished it's propagation. This should help prevent against Let's Encrypt validation failures.

My own testing showed propagation as fast as 20 seconds, and as slow as 3 minutes.